requirements.txt
flask==3.0.0
flask-restful==0.3.10
mysqlclient==2.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
apispec==6.3.1
flask_apispec==0.11.4
PyYAML==6.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0